add animations & resolve conflicts
