+++ frontend/src/components/PostCategories.js
@@ -16,7 +16,8 @@ const PostCategories = props => {
     setShowRestCategories(!showRestCategories)
   }
 
-  const handleClick = (name, id) => {
+  const handleClick = (e, name, id) => {
+    if (e) e.preventDefault()
     history.push(`/category/${name}`)
     dispatch(saveCategoryPosts(id))
   }
@@ -50,7 +51,7 @@ const PostCategories = props => {
       return (
         <SinglePostCategory
           key={id}
-          handleClick={() => handleClick(name, id)}
+          handleClick={e => handleClick(e, name, id)}
           categoryName={name}
           classNames={postSingeCategoryClassNames}
         />
@@ -67,7 +68,7 @@ const PostCategories = props => {
         return (
           <SinglePostCategory
             key={id}
-            handleClick={() => handleClick(name, id)}
+            handleClick={e => handleClick(e, name, id)}
             categoryName={name}
             classNames={postSingeCategoryClassNames}
           />
